+++ app/(afterLogin)/estimate/@estimateTrainer/components/MyEstimateU.tsx
@@ -1,7 +1,8 @@
 "use client";
 
 import { useQuery, useQueryClient } from "@tanstack/react-query";
-import BlankEstimate from "../../estimateUser/components/BlankEstimate";
+
+import BlankEstimate from "../../@estimateUser/components/BlankEstimate";
 import MyEstimateCardU from "./MyEstimateCardU";
 import { sendRequest } from "@/app/api/rootApi";
 import { useUserLocation } from "@/store/useUserLocation";
